v1.8.3 - Redid Permissions (Removed a lot of unnecessary checks) - Updated Packages - Added a 6hr timer for status resync (fixes no status after error bug) - Added top.gg error handler - Updated name in package file

+++ src/Commands/PingCommand.js
@@ -5,8 +5,8 @@ module.exports = {
 		.setName('ping')
 		.setDescription('Gives the ping of Meii!'),
 	async execute(interaction, db, server_data, client) {
+		if(!interaction.channel.permissionsFor(client.user).has(PermissionFlagsBits.SendMessages)) return await interaction.reply({ content: `I'm sorry, I do not have enough permissions to use this command. \nI need.. \`Send Messages\` `, ephemeral: true }).catch(() => {return; })
 		await interaction.deferReply();
-		//Command
 		const m = await interaction.channel.send("Ping?");
 		await interaction.editReply(`:ping_pong: Pong! | Latency is \`${m.createdTimestamp - interaction.createdTimestamp}ms\` | API latency is \`${client.ws.ping}ms\``)
 		m.delete()
